+++ src/main/java/MVC/model/Observer.java
@@ -1,4 +1,4 @@
-package MVC.view;
+package MVC.model;
 
 import MVC.model.Pieces.Piece;
